refactor: move @repo/shared into lib folder, update imports
