+++ apps/app/src/lib/query/_queries/cpalRecorder.ts
@@ -1,6 +1,6 @@
 import { services } from '$lib/services';
 import { toast } from '$lib/toast';
-import type { WhisperingRecordingState } from '@repo/shared';
+import type { WhisperingRecordingState } from '$lib/constants';
 import { defineMutation, defineQuery } from '../_utils';
 import { queryClient } from '../index';
