+++ apps/app/src/lib/query/_queries/download.ts
@@ -2,7 +2,7 @@ import { services } from '$lib/services';
 import type { Recording } from '$lib/services/db';
 import type { DownloadServiceError } from '$lib/services/download/_types';
 import { Err, type Result } from '@epicenterhq/result';
-import type { WhisperingError } from '@repo/shared';
+import type { WhisperingError } from '$lib/result';
 import { defineMutation } from '../_utils';
 
 export const download = {
