+++ apps/app/src/lib/query/_queries/shortcuts.ts
@@ -1,7 +1,7 @@
 import { type Command, type CommandId, commandCallbacks } from '$lib/commands';
 import { services } from '$lib/services';
 import type { Accelerator } from '$lib/services/shortcuts/createGlobalShortcutManager';
-import type { SupportedKey } from '@repo/shared/keyboard';
+import type { SupportedKey } from '$lib/keyboard';
 import { defineMutation } from '../_utils';
 
 export const shortcuts = {
