+++ apps/app/src/lib/query/_queries/sound.ts
@@ -2,7 +2,7 @@ import { services } from '$lib/services';
 import type { PlaySoundServiceError } from '$lib/services/sound/_types';
 import { settings } from '$lib/stores/settings.svelte';
 import { Ok, type Result } from '@epicenterhq/result';
-import type { WhisperingSoundNames } from '@repo/shared';
+import type { WhisperingSoundNames } from '$lib/constants';
 import { defineMutation } from '../_utils';
 
 const soundKeys = {
