+++ apps/app/src/lib/query/_queries/transcription.ts
@@ -4,7 +4,7 @@ import { toast } from '$lib/toast';
 import type { TranscriptionServiceError } from '$lib/services/transcription/_types';
 import { settings } from '$lib/stores/settings.svelte';
 import { Err, Ok, type Result, partitionResults } from '@epicenterhq/result';
-import type { WhisperingError } from '@repo/shared';
+import type { WhisperingError } from '$lib/result';
 import { defineMutation } from '../_utils';
 import { queryClient } from '../index';
 import { recordings } from './recordings';
