+++ apps/app/src/lib/query/_queries/transformer.ts
@@ -4,7 +4,7 @@ import type {
 	TransformationRunFailed,
 } from '$lib/services/db';
 import { Err, Ok, type Result } from '@epicenterhq/result';
-import type { WhisperingError, WhisperingResult } from '@repo/shared';
+import type { WhisperingError, WhisperingResult } from '$lib/result';
 import { defineMutation } from '../_utils';
 import { queryClient } from '../index';
 import { transformationRunKeys } from './transformationRuns';
