+++ apps/app/src/lib/services/clipboard/_types.ts
@@ -3,7 +3,7 @@ import type {
 	MaybePromise,
 	WhisperingError,
 	WhisperingWarning,
-} from '@repo/shared';
+} from '$lib/result';
 
 export type ClipboardServiceError = TaggedError<'ClipboardServiceError'>;
