+++ apps/app/src/lib/services/clipboard/desktop.ts
@@ -3,7 +3,7 @@ import { invoke } from '@tauri-apps/api/core';
 import { writeText } from '@tauri-apps/plugin-clipboard-manager';
 import { type } from '@tauri-apps/plugin-os';
 import type { ClipboardService, ClipboardServiceError } from './_types';
-import type { WhisperingWarning } from '@repo/shared';
+import type { WhisperingWarning } from '$lib/result';
 
 export function createClipboardServiceDesktop(): ClipboardService {
 	return {
