+++ apps/app/src/lib/services/cpalRecorder.ts
@@ -3,7 +3,7 @@ import type { Result, TaggedError } from '@epicenterhq/result';
 import type {
 	WhisperingRecordingState,
 	CancelRecordingResult,
-} from '@repo/shared';
+} from '$lib/constants';
 import { invoke as tauriInvoke } from '@tauri-apps/api/core';
 
 type RecordingServiceError = TaggedError<'RecordingServiceError'>;
