+++ apps/app/src/lib/services/db/models.ts
@@ -5,7 +5,7 @@ import type {
 	GROQ_INFERENCE_MODELS,
 	INFERENCE_PROVIDERS,
 	OPENAI_INFERENCE_MODELS,
-} from '@repo/shared';
+} from '$lib/constants';
 import { nanoid } from 'nanoid/non-secure';
 
 export type DbServiceErrorProperties = TaggedError<'DbServiceError'>;
