+++ apps/app/src/lib/services/transcription/whisper/openai.ts
@@ -1,5 +1,5 @@
 import { Err, Ok } from '@epicenterhq/result';
-import type { WhisperingError } from '@repo/shared';
+import type { WhisperingError } from '$lib/result';
 import type { TranscriptionService } from '../_types';
 import type { HttpService } from '../../http/_types';
 import { createWhisperService } from './_createWhisperService';
