+++ apps/app/src/lib/settings/index.ts
@@ -1,0 +1,10 @@
+export * from './settingsV1';
+export * from './settingsV2';
+export * from './settingsV3';
+export * from './settingsV4';
+export * from './settingsV5';
+export * from './settingsV6';
+export * from './settingsV7';
+
+export type { SettingsV7 as Settings } from './settingsV7';
+export { settingsV7Schema as settingsSchema } from './settingsV7';
