+++ apps/app/src/routes/(config)/settings/+page.svelte
@@ -8,7 +8,7 @@
 	import { Button } from '$lib/components/ui/button/index.js';
 	import { Separator } from '$lib/components/ui/separator/index.js';
 	import { settings } from '$lib/stores/settings.svelte';
-	import { ALWAYS_ON_TOP_OPTIONS } from '@repo/shared';
+	import { ALWAYS_ON_TOP_OPTIONS } from '$lib/constants';
 	import { type } from '@tauri-apps/plugin-os';
 </script>
 
@@ -79,7 +79,6 @@
 	/>
 
 	<Separator />
-
 
 	{#if window.__TAURI_INTERNALS__}
 		<LabeledSwitch
