+++ apps/app/src/routes/(config)/settings/shortcuts/keyboard-shortcut-recorder/GlobalKeyboardShortcutRecorder.svelte
@@ -7,7 +7,7 @@
 		pressedKeysToTauriAccelerator,
 		type Accelerator,
 	} from '$lib/services/shortcuts/createGlobalShortcutManager';
-	import type { SupportedKey } from '@repo/shared/keyboard';
+	import type { SupportedKey } from '$lib/keyboard';
 	import { settings } from '$lib/stores/settings.svelte';
 	import { toast } from '$lib/toast';
 	import { type PressedKeys } from '$lib/utils/createPressedKeys.svelte';
