+++ apps/app/src/routes/(config)/settings/shortcuts/keyboard-shortcut-recorder/create-key-recorder.svelte.ts
@@ -1,4 +1,4 @@
-import type { SupportedKey } from '@repo/shared/keyboard';
+import type { SupportedKey } from '$lib/keyboard';
 import type { PressedKeys } from '$lib/utils/createPressedKeys.svelte';
 
 const CAPTURE_WINDOW_MS = 300; // Time to wait for additional keys in a combination
