--- packages/shared/src/settings/index.ts
@@ -1,10 +1,0 @@
-export * from './settingsV1.js';
-export * from './settingsV2.js';
-export * from './settingsV3.js';
-export * from './settingsV4.js';
-export * from './settingsV5.js';
-export * from './settingsV6.js';
-export * from './settingsV7.js';
-
-export type { SettingsV7 as Settings } from './settingsV7.js';
-export { settingsV7Schema as settingsSchema } from './settingsV7.js';
